requirements.txt
openai
python-dotenv
sounddevice
soundfile
numpy
pydub
edge-tts
faster-whisper
simpleaudio